+++ tests/testthat/test-get.R
@@ -327,3 +327,41 @@ test_that("get_packages() works as expected with `content_guid` names in API res
     )
   )
 })
+
+test_that("get_content only requests vanity URLs for Connect 2024.06.0 and up", {
+  with_mock_dir("2024.05.0", {
+    client <- Connect$new(server = "http://connect.example", api_key = "not-a-key")
+    # `$version` is lazy, so we need to call it before `without_internet()`.
+    client$version
+  })
+  without_internet({
+    expect_GET(
+      get_content(client),
+      "http://connect.example/__api__/v1/content?include=tags%2Cowner"
+    )
+  })
+
+  with_mock_dir("2024.06.0", {
+    client <- Connect$new(server = "http://connect.example", api_key = "not-a-key")
+    # `$version` is lazy, so we need to call it before `without_internet()`.
+    client$version
+  })
+  without_internet({
+    expect_GET(
+      get_content(client),
+      "http://connect.example/__api__/v1/content?include=tags%2Cowner"
+    )
+  })
+
+  with_mock_dir("2024.07.0", {
+    client <- Connect$new(server = "http://connect.example", api_key = "not-a-key")
+    # `$version` is lazy, so we need to call it before `without_internet()`.
+    client$version
+  })
+  without_internet({
+    expect_GET(
+      get_content(client),
+      "http://connect.example/__api__/v1/content?include=tags%2Cowner%2Cvanity_url"
+    )
+  })
+})